Clean up dataset and add comparison plots
